move SanitizeError to sql helper lib
