+++ x-pack/metricbeat/module/sql/query/dsn.go
@@ -14,6 +14,7 @@ import (
 	"github.com/godror/godror"
 	"github.com/godror/godror/dsn"
 
+	"github.com/elastic/beats/v7/metricbeat/helper/sql"
 	"github.com/elastic/beats/v7/metricbeat/mb"
 	"github.com/elastic/elastic-agent-libs/logp"
 	"github.com/elastic/elastic-agent-libs/transport/tlscommon"
@@ -31,11 +32,11 @@ type ConnectionDetails struct {
 const mysqlTLSConfigKey = "custom"
 
 // ParseDSN tries to parse the host
-// ParseDSN may return an error: it is a sanitized error recreated with errors.New (via defer sanitizeError) to avoid leaking sensitive data.
+// ParseDSN may return an error: it is a sanitized error recreated with errors.New (via defer sql.SanitizeError) to avoid leaking sensitive data.
 // The returned error does not wrap the original, so errors.Is/As will not match.
 func ParseDSN(mod mb.Module, host string) (_ mb.HostData, fetchErr error) {
 	defer func() {
-		fetchErr = sanitizeError(fetchErr, host)
+		fetchErr = sql.SanitizeError(fetchErr, host)
 	}()
 
 	logger := logp.NewLogger("")
